requirements.txt
vectors==1.0.0
pyudev==0.21.0
pyserial==3.4
Twisted==17.9.0
zeroconf==0.19.1
